chore(ui): revert changes
