+++ ui/src/components/flows/FlowMetrics.vue
@@ -23,11 +23,6 @@
             settings: {shown: false}
         }"
     />
-     <!-- Add the ScopeFilterButton component here -->
-        <scope-filter-button
-            label="Metric"
-            @update:modelValue="handleScopeUpdate"
-        />
 
     <div v-bind="$attrs" v-loading="isLoading">
         <el-card>
@@ -68,14 +63,12 @@
     import {defaultConfig, getFormat, tooltip} from "../../utils/charts";
     import {cssVariable} from "@kestra-io/ui-libs/src/utils/global";
     import KestraFilter from "../filter/KestraFilter.vue";
-    import ScopeFilterButton from './ScopeFilterButton.vue';
 
     export default {
         name: "FlowMetrics",
         components: {
             Bar,
             KestraFilter,
-            ScopeFilterButton,
         },
         created() {
             this.loadMetrics();
@@ -272,9 +265,7 @@
                 } else {
                     this.loadAggregatedMetrics();
                 }
-            },
-            handleScopeUpdate(value) {
-            },
+            }
         },
         watch: {
             "$route.query": {
