+++ ui/src/components/flows/FlowRoot.vue
@@ -7,11 +7,6 @@
             ref="currentTab"
             :tabs="tabs"
         />
-        <!-- Add the ScopeFilterButton component here -->
-        <scope-filter-button
-            label="Flow"
-            @update:modelValue="handleScopeUpdate"
-        />
     </template>
 </template>
 
@@ -34,24 +29,12 @@
     import {apiUrl} from "override/utils/route";
     import FlowRootTopBar from "./FlowRootTopBar.vue";
     import FlowConcurrency from "./FlowConcurrency.vue";
-    import ScopeFilterButton from './ScopeFilterButton.vue';
 
     export default {
         mixins: [RouteContext],
         components: {
             Tabs,
-            Topology,
-            FlowRevisions,
-            LogsWrapper,
-            FlowExecutions,
-            Overview,
-            FlowDependencies,
-            FlowNoDependencies,
-            FlowMetrics,
-            FlowEditor,
-            FlowTriggers,
             FlowRootTopBar,
-            ScopeFilterButton,
         },
         data() {
             return {
@@ -299,15 +282,12 @@
             },
             activeTabName() {
                 return this.$refs.currentTab?.activeTab?.name ?? "home";
-            },
-            handleScopeUpdate(value) {
-            },
+            }
         },
         computed: {
             ...mapState("flow", ["flow"]),
             ...mapState("auth", ["user"]),
             ...mapState("core", ["guidedProperties"]),
-            ...mapState("flow", ["ready", "routeInfo", "deleted", "isAllowedEdit"]),
             routeInfo() {
                 return {
                     title: this.$route.params.id,
